+++ dart/geobase/lib/tiling.dart
@@ -11,5 +11,6 @@
 /// Usage: import `package:geobase/tiling.dart`
 library tiling;
 
+export 'src/constants/screen_ppi.dart';
 export 'src/tiling/convert.dart';
 export 'src/tiling/tilematrix.dart';